refactor: move ar-io-peer-manager to peers directory
- Create new src/peers/ directory for peer management code
- Move ArIOPeerManager and its test from src/data/ to src/peers/
- Update all import paths to reflect new location
- No functionality changes, pure refactoring for better code organization
